requirements.txt
beautifulsoup4==4.6.3
celery==4.2.1
flower==0.9.2
halo==0.0.22
requests==2.20.1
redis==3.0.1
pybind11==2.2.4
python-mecab-ko==1.0.4
uvloop==0.11.3
